+++ docs/changelog.rst
@@ -11,6 +11,7 @@ Changelog
   This feature is also available as a hook in ``schemathesis.contrib.unique_data``.
 - ``schemathesis.checks.register`` as a new way to register checks.
 - ``schemathesis.targets.register`` as a new way to register targets.
+- ``schemathesis.openapi.format`` as a new way to register custom formats for Open API.
 
 **Changed**
 
@@ -23,6 +24,7 @@ Changelog
 
 - ``schemathesis.register_check`` in favor of ``schemathesis.checks.register``.
 - ``schemathesis.register_target`` in favor of ``schemathesis.targets.register``.
+- ``schemathesis.register_string_format`` in favor of ``schemathesis.openapi.format``.
 
 **Fixed**